# dto/response.go
package dto

type LoginResponse struct {
	Token string `json:"token"`
}

type ResponseFailed struct {
	Message string `json:"message"`
}

type ResponseFailedUnauthorized struct {
	Message string `json:"message"`
}